+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 - Corrected the Latvian translation of the phrase "Select terminal"
 
+### Improved
+- added a ability to disable Omniva shipping methods for specific shipping class
+
 ## [1.11.1]
 ### Fixed
 - fixed error, when cart product dont have category (product category is "Uncategorized")